corrections

+++ src/main/java/ru/netology/service/PostService.java
@@ -28,6 +28,10 @@ public class PostService {
   }
 
   public void removeById(long id) {
-    repository.removeById(id);
+    if (repository.getById(id) != null) {
+      repository.removeById(id);
+    } else {
+      throw new NotFoundException();
+    }
   }
 }
